[!!!][TASK] Change configuration directory
Limiting the configuration file folder to a specific folder path
makes it possible to introduce some clarity about the extension.

If you use the command without providing a path to a valid yaml file,
Whenever you don't provide a path to a yaml file while calling the
command, all active packages (extensions) are checked for valid
yaml files in EXT:yourExtension/Configuration/YamlConfiguration/

This patch updates the path and the documentation of this existing
feature.

Fixes: #33